api/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
tubular
lightgbm
PyYAML==6.0
python-dotenv==1.0.0
azure-storage-blob==12.12.0
azure-keyvault-secrets==4.7.0
azure-identity==1.12.0